requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
appdirs==1.4.4
APScheduler==3.10.4
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
binteger==0.8.0
bitarray==2.9.2
black==24.4.2
bleach==6.2.0
blinker==1.8.2
capstone==5.0.1
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colored-traceback==0.4.2
comm==0.2.2
constantly==23.10.4
cry==0.2.4
cryptography==43.0.1
cssselect==1.2.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
discord.py==2.4.0
dnspython==2.6.1
docopt==0.6.2
eventlet==0.36.1
executing==2.0.1
factordb-pycli==1.3.0
fastjsonschema==2.20.0
filebytes==0.10.2
filelock==3.16.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
frida==16.5.4
frida-tools==13.3.0
frozenlist==1.4.1
gmpy2==2.2.1
greenlet==3.1.1
gunicorn==22.0.0
hashpumpy==1.2
hyperlink==21.0.0
idna==2.10
importlib_metadata==8.4.0
incremental==24.7.2
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.26.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwt==1.3.1
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.1
oscrypto==1.3.0
packaging==24.1
padding_oracle==0.4.1
pandocfilters==1.5.1
paramiko==3.4.0
parsel==1.9.1
parso==0.8.4
pathspec==0.12.1
pdf2john==0.2.0
pexpect==4.9.0
pikepdf==9.2.1
pillow==10.4.0
platformdirs==4.2.2
plumbum==1.8.3
prompt_toolkit==3.0.47
Protego==0.3.1
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
pwntools==4.13.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.10.4
pycurl==7.45.3
PyDispatcher==2.0.7
pyelftools==0.31
Pygments==2.18.0
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pyjokes==0.8.3
PyJWT==2.9.0
pymongo==4.8.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
pyserial==3.5
pyshark==0.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.0.3
qrcode==7.4.2
queuelib==1.7.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
requests-file==2.1.0
ROPGadget==7.4
ropper==1.13.8
rpds-py==0.21.0
rpyc==6.0.0
Scrapy==2.11.2
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
termcolor==2.4.0
tinycss2==1.4.0
tk==0.1.0
tldextract==5.1.2
tokenize-rt==5.2.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
Twisted==24.7.0
typing_extensions==4.12.2
tzlocal==5.2
unicorn==2.0.1.post1
unix-ar==0.2.1
uritools==4.0.3
urllib3==1.26.5
w3lib==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
xortool==1.0.2
yarl==1.13.1
z3-solver==4.13.0.0
zipp==3.20.1
zope.interface==7.1.0
zstandard==0.23.0
